pyproject.toml
[project]
name = "discord-attachment-downloader"
version = "0.0.1"
dependencies = [
    "ffmpeg-python",
    "requests"
]